+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -13,6 +13,8 @@
   ([#42551](https://github.com/Azure/azure-sdk-for-python/pull/42551))
 - Rename Customer Statsbeat to Customer SDKStats as per [Spec] - https://github.com/aep-health-and-standards/Telemetry-Collection-Spec/pull/581
   ([#42573](https://github.com/Azure/azure-sdk-for-python/pull/42573))
+- Customer Facing SDKStats: Exception categorization as per [Spec] - https://github.com/aep-health-and-standards/Telemetry-Collection-Spec/blob/main/ApplicationInsights/sdkstats/customer_facing_sdk_stats.md
+  ([#42695](https://github.com/Azure/azure-sdk-for-python/pull/42695))
 - Customer Facing SDKStats: Added logic for race conditions and updated the implementation to use a global instance for customer SDKStats metrics
   ([#42655](https://github.com/Azure/azure-sdk-for-python/pull/42655))
 ### Breaking Changes